initial movement for mint
