+++ x/mint/genesis.go
@@ -1,0 +1,1 @@
+package mint
