+++ x/mint/minter.go
@@ -1,0 +1,52 @@
+package mint
+
+import "time"
+
+// Pool - dynamic parameters of the current state
+type Minter struct {
+	InflationLastTime time.Time `json:"inflation_last_time"` // block which the last inflation was processed
+	Inflation         sdk.Dec   `json:"inflation"`           // current annual inflation rate
+}
+
+const precision = 10000            // increased to this precision for accuracy
+var hrsPerYrDec = sdk.NewDec(8766) // as defined by a julian year of 365.25 days
+
+// process provisions for an hour period
+func (m Minter) ProcessProvisions(params Params) Pool {
+	p.Inflation = p.NextInflation(params)
+	provisions := p.Inflation.
+		Mul(p.TokenSupply()).
+		Quo(hrsPerYrDec)
+
+	// TODO add to the fees provisions
+	p.LooseTokens = p.LooseTokens.Add(provisions)
+	return p
+}
+
+// get the next inflation rate for the hour
+func (m Minter) NextInflation(params Params) (inflation sdk.Dec) {
+
+	// The target annual inflation rate is recalculated for each previsions cycle. The
+	// inflation is also subject to a rate change (positive or negative) depending on
+	// the distance from the desired ratio (67%). The maximum rate change possible is
+	// defined to be 13% per year, however the annual inflation is capped as between
+	// 7% and 20%.
+
+	// (1 - bondedRatio/GoalBonded) * InflationRateChange
+	inflationRateChangePerYear := sdk.OneDec().
+		Sub(p.BondedRatio().
+			Quo(params.GoalBonded)).
+		Mul(params.InflationRateChange)
+	inflationRateChange := inflationRateChangePerYear.Quo(hrsPerYrDec)
+
+	// increase the new annual inflation for this next cycle
+	inflation = p.Inflation.Add(inflationRateChange)
+	if inflation.GT(params.InflationMax) {
+		inflation = params.InflationMax
+	}
+	if inflation.LT(params.InflationMin) {
+		inflation = params.InflationMin
+	}
+
+	return inflation
+}
