+++ x/stake/types/pool.go
@@ -11,15 +11,9 @@ import (
 
 // Pool - dynamic parameters of the current state
 type Pool struct {
-	LooseTokens       sdk.Dec   `json:"loose_tokens"`        // tokens which are not bonded in a validator
-	BondedTokens      sdk.Dec   `json:"bonded_tokens"`       // reserve of bonded tokens
-	InflationLastTime time.Time `json:"inflation_last_time"` // block which the last inflation was processed
-	Inflation         sdk.Dec   `json:"inflation"`           // current annual inflation rate
-
-	DateLastCommissionReset int64 `json:"date_last_commission_reset"` // unix timestamp for last commission accounting reset (daily)
-
-	// Fee Related
-	PrevBondedShares sdk.Dec `json:"prev_bonded_shares"` // last recorded bonded shares - for fee calculations
+	LooseTokens             sdk.Dec `json:"loose_tokens"`               // tokens which are not bonded in a validator
+	BondedTokens            sdk.Dec `json:"bonded_tokens"`              // reserve of bonded tokens
+	DateLastCommissionReset int64   `json:"date_last_commission_reset"` // unix timestamp for last commission accounting reset (daily)
 }
 
 // nolint
@@ -37,7 +31,6 @@ func InitialPool() Pool {
 		InflationLastTime:       time.Unix(0, 0),
 		Inflation:               sdk.NewDecWithPrec(7, 2),
 		DateLastCommissionReset: 0,
-		PrevBondedShares:        sdk.ZeroDec(),
 	}
 }
 
@@ -79,52 +72,6 @@ func (p Pool) bondedTokensToLoose(bondedTokens sdk.Dec) Pool {
 	return p
 }
 
-//_______________________________________________________________________
-// Inflation
-
-const precision = 10000            // increased to this precision for accuracy
-var hrsPerYrDec = sdk.NewDec(8766) // as defined by a julian year of 365.25 days
-
-// process provisions for an hour period
-func (p Pool) ProcessProvisions(params Params) Pool {
-	p.Inflation = p.NextInflation(params)
-	provisions := p.Inflation.
-		Mul(p.TokenSupply()).
-		Quo(hrsPerYrDec)
-
-	// TODO add to the fees provisions
-	p.LooseTokens = p.LooseTokens.Add(provisions)
-	return p
-}
-
-// get the next inflation rate for the hour
-func (p Pool) NextInflation(params Params) (inflation sdk.Dec) {
-
-	// The target annual inflation rate is recalculated for each previsions cycle. The
-	// inflation is also subject to a rate change (positive or negative) depending on
-	// the distance from the desired ratio (67%). The maximum rate change possible is
-	// defined to be 13% per year, however the annual inflation is capped as between
-	// 7% and 20%.
-
-	// (1 - bondedRatio/GoalBonded) * InflationRateChange
-	inflationRateChangePerYear := sdk.OneDec().
-		Sub(p.BondedRatio().
-			Quo(params.GoalBonded)).
-		Mul(params.InflationRateChange)
-	inflationRateChange := inflationRateChangePerYear.Quo(hrsPerYrDec)
-
-	// increase the new annual inflation for this next cycle
-	inflation = p.Inflation.Add(inflationRateChange)
-	if inflation.GT(params.InflationMax) {
-		inflation = params.InflationMax
-	}
-	if inflation.LT(params.InflationMin) {
-		inflation = params.InflationMin
-	}
-
-	return inflation
-}
-
 // HumanReadableString returns a human readable string representation of a
 // pool.
 func (p Pool) HumanReadableString() string {
@@ -134,10 +81,7 @@ func (p Pool) HumanReadableString() string {
 	resp += fmt.Sprintf("Bonded Tokens: %s\n", p.BondedTokens)
 	resp += fmt.Sprintf("Token Supply: %s\n", p.TokenSupply())
 	resp += fmt.Sprintf("Bonded Ratio: %v\n", p.BondedRatio())
-	resp += fmt.Sprintf("Previous Inflation Block: %s\n", p.InflationLastTime)
-	resp += fmt.Sprintf("Inflation: %v\n", p.Inflation)
 	resp += fmt.Sprintf("Date of Last Commission Reset: %d\n", p.DateLastCommissionReset)
-	resp += fmt.Sprintf("Previous Bonded Shares: %v\n", p.PrevBondedShares)
 	return resp
 }
